% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{set_bottom_text.magick-image}
\alias{set_bottom_text.magick-image}
\title{Creates title-like text at the bottom of an image.}
\usage{
\method{set_bottom_text}{magick-image}(i, text, gravity = "northwest",
  location = "0+0", degrees = 0, height = 10, size = 10,
  font = "", color = NULL, bg_color = "none", strokecolor = NULL,
  boxcolor = NULL)
}
\description{
Creates title-like text at the bottom of an image.
}
