% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{set_top_text.magick-image}
\alias{set_top_text.magick-image}
\title{Creates title-like text at the top of an image.}
\usage{
\method{set_top_text}{magick-image}(i, text, gravity = "northwest",
  location = "0+0", degrees = 0, height = 10, size = 10,
  font = "", color = NULL, bg_color = "none", strokecolor = NULL,
  boxcolor = NULL)
}
\arguments{
\item{height}{Height of the text image given in pixels (numeric) or percent
(character).}
}
\description{
Creates title-like text at the top of an image.
}
